+++ src/test/run-pass/expr-copy.rs
@@ -15,8 +15,8 @@ fn f(arg: &mut A) {
     arg.a = 100;
 }
 
-#[derive(Copy)]
-struct A { a: int }
+#[derive(Copy, Clone)]
+struct A { a: isize }
 
 pub fn main() {
     let mut x = A {a: 10};
